infra/cdk/requirements.txt
aws-cdk-lib>=2.130.0
aws-cdk.lambda-layer-kubectl-v30>=2.0.0
constructs>=10.3.0
boto3>=1.34.0
